report-api/requirements.txt
Brotli==1.1.0
Flask==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
cairocffi==1.7.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cssselect2==0.7.0
flask-restx==1.3.0
flask_weasyprint==1.1.0
fonttools==4.53.0
gunicorn==22.0.0
html5lib==1.1
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
jaeger-client==4.8.0
jsonschema==4.17.3
opentracing==2.4.0
packaging==24.0
pillow==10.3.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.10.0
pyphen==0.15.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
threadloop==1.0.2
thrift==0.20.0
tinycss2==1.3.0
tornado==6.4
urllib3==2.2.2
weasyprint==61.2
webencodings==0.5.1
zipp==3.19.1
zopfli==0.2.3
-e git+https://github.com/bcgov/sbc-common-components.git#egg=sbc-common-components&subdirectory=python
git+https://github.com/daxiom/simple-cloudevent.py.git
git+https://github.com/seeker25/flask-jwt-oidc.git
pikepdf==9.10.2
psutil==5.9.8
aiohttp==3.12.15
